Add a public portability API for `*Raw*`.
`RawFilelike` et al are portability abstractions over `RawFd` and
`RawHandle`/`RawSocket` and their associated traits.

Define these as public types and traits in a new public `raw` module.
